Split the cycle detection to a new file.
Make also the interface nicer.
